feat: Optimistics updates have been added to the workspace loader screen
Eventually it was observed that loading a workspace was an heavier tasks and it was taking quite a decent amount of time, and for that time being we were showing just  which was giving some short hang state feeling to the user.

Now this update that message is updating in every 3 seconds, majorily there 4 message for now and the very last message is  which will persists to the very end of the loading state.

Thanks