feat: Optimistic updates have been added in the workspace
Optimistic updates for the saving the data into db has been added into the Navbar of the applications workspace area. This will provide user of dynamically uploading data into the db

The default value for optimistics update notification has been set to 3000ms that is 3 seconds. This can be changed at any point of the application

Thanks